+++ app/models/character.go
@@ -1,0 +1,8 @@
+package app
+
+import "go-rpg/proto"
+
+type Character struct {
+	id int32
+	proto.Character
+}
