+++ setup/db.go
@@ -1,0 +1,3 @@
+package setup
+
+type Database interface{}
